refactored request_queue to separate file
